_requirements.txt
bidict==0.21.2
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-engineio==4.1.0
python-socketio==5.2.1
Werkzeug==1.0.1
eventlet==0.24.1
